backend/requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy
psycopg2-binary
alembic

redis
geoalchemy2

pydantic-settings
python-dotenv
asyncpg

openai
requests
pandas
websockets
httpx
# fastapi-websockets

aiohttp
tweepy

streamlit 
streamlit-leaflet 
